requirements.txt
aiosqlite==0.20.0
alembic==1.13.2
annotated-types==0.7.0
anthropic==0.39.0
anyio==4.7.0
certifi==2024.8.30
charset-normalizer==3.4.0
distro==1.9.0
greenlet==3.1.1
groq==0.6.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.0
idna==3.10
Jinja2==3.1.4
jiter==0.8.0
jsonref==1.1.0
Mako==1.3.7
MarkupSafe==3.0.2
openai==1.40.6
prompt-toolkit==3.0.48
psutil==5.9.8
pydantic-core==2.20.1
pydantic==2.8.2
python-dotenv==1.0.1
pytest-cov==7.0.0
pytest-timeout==2.4.0
regex==2024.11.6
requests==2.32.3
sniffio==1.3.1
sqlalchemy==2.0.32
sqlalchemy[asyncio]==2.0.32
tenacity==9.0.0
tiktoken==0.8.0
trafilatura==1.6.4
tqdm==4.67.1
typing-extensions==4.12.2
urllib3==2.2.3
wcwidth==0.2.13
asyncpg==0.29.0
psycopg2-binary==2.9.9
pgvector==0.2.5
